ENH: Adding an API for interpolation functions and derivatives that makes it possible to implement for cell like vtkPolygon. For all other cell makeing the virtual function call the static one (for now).

+++ VTK/Common/vtkWin32ProcessOutputWindow.cxx
@@ -148,7 +148,7 @@ int vtkWin32ProcessOutputWindow::Initialize()
 }
 
 //----------------------------------------------------------------------------
-void vtkWin32ProcessOutputWindow::Write(const char* data, int length)
+void vtkWin32ProcessOutputWindow::Write(const char* data, size_t length)
 {
   if(data && length)
     {
